Fix build --install failure on SDK level (#71)
On SDK(top) level, build.py --install was failing, due to the logic on how
ML_SDK_GENERATE_CPACK option was handled on top and component levels.
In this repo, we need to pass this option for the --install dir as well
to allow loading the install rules from the package.cmake file here,
but make sure we are including the package rules within the component cmake configs.


Change-Id: I3082653c12026f026edddeafe6b6e6beb768de6f

Signed-off-by: Gergely Nagy <gergely.nagy@arm.com>